+++ CHANGELOG.md
@@ -9,6 +9,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- ✨ **changelog**: Added automatic changelog system
+
+
+## [0.2.0] - 2025-08-07
+
+### Feature Enhancement
+
+This minor release includes 11 changes with 5 new features, 2 bug fixes, 4 improvements. Enjoy the new features and enhanced functionality!
+
+### Added
+
 - Changelog app to view project updates
 - Multiple file tabs support in Markdown Editor
 - File tree sidebar with search functionality